refactor: move Permit type to separate module with documentation
